[REFECT] 집 등록 및 조회기능 ref
