+++ src/main/java/konkuk/tourkk/chons/domain/sigungu/application/res/SigunguResponse.java
@@ -1,4 +1,4 @@
-package konkuk.tourkk.chons.domain.festival.application.res;
+package konkuk.tourkk.chons.domain.sigungu.application.res;
 
 import lombok.Getter;
