+++ src/main/java/konkuk/tourkk/chons/domain/sigungu/infrastructure/SigunguRepository.java
@@ -1,7 +1,7 @@
-package konkuk.tourkk.chons.domain.festival.infrastructure;
+package konkuk.tourkk.chons.domain.sigungu.infrastructure;
 
 import java.util.Optional;
-import konkuk.tourkk.chons.domain.festival.domain.entity.Sigungu;
+import konkuk.tourkk.chons.domain.sigungu.domain.entity.Sigungu;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface SigunguRepository extends JpaRepository<Sigungu, Long> {
