+++ src/main/java/konkuk/tourkk/chons/global/common/webclient/WebClientService.java
@@ -6,8 +6,7 @@ import java.net.URI;
 import java.time.LocalDate;
 import java.util.Map;
 import konkuk.tourkk.chons.domain.festival.domain.entity.Area;
-import konkuk.tourkk.chons.domain.festival.domain.entity.Sigungu;
-import lombok.extern.slf4j.Slf4j;
+import konkuk.tourkk.chons.domain.sigungu.domain.entity.Sigungu;
 import org.springframework.stereotype.Component;
 import org.springframework.web.reactive.function.client.WebClient;
 import org.springframework.web.util.UriBuilder;
